Refactor mantle errors
