+++ mantle/src/exe.rs
@@ -93,17 +93,39 @@ impl Context {
     }
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub enum RpcError<T> {
     /// There was no service at the requested address.
-    NoService,
+    NoAccount,
 
     /// The caller does not have enough balance to cover the sent value.
-    InsufficientBalance,
+    InsufficientFunds,
 
     /// The caller did not provide enough gas to complete the transaction.
     InsufficientGas,
 
+    InvalidInput,
+
+    InvalidOutput(Vec<u8>),
+
     /// The application returned an error.
     Exec(T),
+}
+
+impl<T: serde::de::DeserializeOwned> From<crate::backend::Error> for RpcError<T> {
+    fn from(err: crate::backend::Error) -> Self {
+        use crate::backend::Error as BackendError;
+        match err {
+            BackendError::Unknown => panic!("Unknown error occured."),
+            BackendError::InsufficientFunds => RpcError::InsufficientFunds,
+            BackendError::InvalidInput => RpcError::InvalidInput,
+            BackendError::NoAccount => RpcError::NoAccount,
+            BackendError::Execution { payload, .. } => {
+                RpcError::Exec(match serde_cbor::from_slice::<T>(&payload) {
+                    Ok(t) => t,
+                    Err(_) => return RpcError::InvalidOutput(payload),
+                })
+            }
+        }
+    }
 }
